src/bff_web/requirements.txt
python-dotenv
requests
fastapi==0.89.1
uvicorn[standard]==0.20.0
